Extract common code to RequestUtilities files.
